requirements.txt
gradio==3.24.1
matplotlib==3.7.1
networkx==3.0
numpy==1.23.5
opencv_contrib_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==1.13.1
torchvision==0.14.1
